requirements.txt
scrapegraphai>=1.13.0
playwright>=1.44.0
python-dotenv>=1.0.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.29.0
